#27 - Service Interface 도입, AbstractService 도입
* 주석과 코드의 분리를 위해 Service Interface를 추가했습니다
* 에러 던질 때에 ServiceType을 넘겨 주는 것이 번거로우므로 `AbstractService` 를 통해 해당 서비스의 ServiceType을 넘기도록 했습니다.
